Add a way to load older episodes

+++ shared/src/commonMain/kotlin/com/ramitsuri/podcasts/network/model/GetEpisodesRequest.kt
@@ -3,5 +3,5 @@ package com.ramitsuri.podcasts.network.model
 internal data class GetEpisodesRequest(
     val id: Long,
     val sinceEpochSeconds: Long? = null,
-    val max: Int = 50,
+    val max: Long = 50,
 )
